Ensure no integer overflow happens.

+++ core/src/main/java/com/github/weisj/darklaf/graphics/Animator.java
@@ -120,7 +120,7 @@ public abstract class Animator {
                         });
                     }
                 }
-            }, 0, cycleDuration * 1000 / totalFrames, TimeUnit.MICROSECONDS);
+            }, 0, cycleDuration * 1000L / totalFrames, TimeUnit.MICROSECONDS);
         }
     }
 
@@ -142,7 +142,7 @@ public abstract class Animator {
 
         if (startTime == -1) {
             startTime = System.currentTimeMillis();
-            stopTime = startTime + (cycleDuration * (totalFrames - currentFrame)) / totalFrames;
+            stopTime = startTime + (((long) cycleDuration) * (totalFrames - currentFrame)) / totalFrames;
         }
 
         final double passedTime = System.currentTimeMillis() - startTime;
